fix: rename featured-cards.demo.tsx file
