+++ components/application/app-navigation/base-components/featured-cards.tsx
