+++ components/application/app-navigation/sidebar-navigation-base.tsx
@@ -1,4 +1,4 @@
-export * as FeaturedCardDemo from "./base-components/featured-cards.demo";
+export * as FeaturedCardDemo from "./base-components/featured-cards";
 export { MobileNavigationHeader } from "./base-components/mobile-header";
 export { NavAccountCard } from "./base-components/nav-account-card";
 export { NavItemButton } from "./base-components/nav-item-button";
